refactor: post디렉토리에서 vote로 경로 변경
